requirements.txt
# Dépendances pour le Bot de Trading Crypto Avancé
# ===============================================

# API Kraken
krakenex==2.1.0
pykrakenapi==0.3.1

# Analyse de données
pandas>=1.5.0
numpy>=1.21.0

# Indicateurs techniques
ta-lib>=0.4.24
scipy>=1.9.0

# Planification et timing
schedule>=1.2.0

# Configuration et variables d'environnement
python-dotenv>=0.19.0

# Logging et monitoring
colorlog>=6.6.0

# Interface utilisateur (optionnel)
streamlit>=1.22.0
plotly>=5.10.0

# Notifications (optionnel)
requests>=2.28.0

# Tests
pytest>=7.0.0
pytest-mock>=3.8.0

# Développement
black>=22.0.0
flake8>=4.0.0
mypy>=0.950 